requirements.txt
tensorflow==1.14.0
numpy==1.16.2
Flask==1.0.2
Pillow==8.4.0
